[WIP] Refactors models
